fix: more things

+++ crates/analyzer/src/err.rs
@@ -77,7 +77,7 @@ pub enum ErrorKind {
     IncompatibleTypes(Type, Type),
     ExpectedArray(Type),
     UnknownMemberFunction,
-    WrongNumberOfTypeArguments(usize, i32),
+    WrongNumberOfTypeArguments(usize, usize),
     OnlyVariablesAllowedInEnumPattern,
     ExpressionsNotAllowedInLetPattern,
     UnknownField,
